+++ lib/src/input.dart
@@ -8,8 +8,10 @@ import 'package:intl_phone_field/countries.dart';
 import 'package:intl_phone_field/intl_phone_field.dart';
 import 'package:intl_phone_field/phone_number.dart';
 import 'package:logger/logger.dart';
+import 'package:sdui/sdui.dart';
 
 import 'action.dart';
+import 'button.dart';
 import 'form.dart';
 import 'widget.dart';
 
@@ -214,11 +216,10 @@ class _SubmitWidgetState extends State<_SubmitWidgetStateful> {
   }
 
   @override
-  Widget build(BuildContext context) => SizedBox(
-      width: double.maxFinite,
-      child: ElevatedButton(
-          child: Text(delegate.caption ?? "Submit"),
-          onPressed: () => _onSubmit(context)));
+  Widget build(BuildContext context) => SDUIButton.createWidget(
+      caption: delegate.caption ?? 'Submit',
+      context: context,
+      onPressed: () => _onSubmit(context));
 
   void _onSubmit(BuildContext context) {
     if (!enabled) {